+++ demo/cypress/index.ts
@@ -2,6 +2,7 @@ export { AutoComplete } from './AutoComplete';
 export { Button } from './Button';
 export { ButtonGroup } from './ButtonGroup';
 export { CheckBox } from './CheckBox';
+export { CheckBoxTree } from './CheckBoxTree';
 export { Currency } from './Currency';
 export { Collapse } from './Collapse';
 export { CountDown } from './CountDown';