+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 - DropDown: Added class disabled
+- CheckBoxTree: Added new component
 
 ### Fixed
 - NumericTextBox: Fixed rounding numbers on change event